# lib/api-zod/src/generated/types/createAssignmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Order Management System
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAssignmentRequest {
  orderId: number;
  vehicleId: number;
  driverId: number;
}